move visitor, rename visitor, other changes suggested
